+++ src/jobs/RefreshCacheJob.php
@@ -13,13 +13,11 @@ use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\helpers\ElementTypeHelper;
 use putyourlightson\blitz\helpers\RefreshCacheHelper;
 use putyourlightson\blitz\helpers\SiteUriHelper;
+use putyourlightson\blitz\models\BaseDataModel;
 use putyourlightson\blitz\models\RefreshDataModel;
 use yii\queue\Queue;
 use yii\queue\RetryableJobInterface;
 
-/**
- * TODO: make this job batchable.
- */
 class RefreshCacheJob extends BaseJob implements RetryableJobInterface
 {
     /**
@@ -75,10 +73,18 @@ class RefreshCacheJob extends BaseJob implements RetryableJobInterface
 
         // Merge in site URIs of element IDs to ensure that uncached elements are also generated
         foreach ($refreshData->getElementTypes() as $elementType) {
-            /** @var ElementInterface|string $elementType */
+            /** @var class-string<ElementInterface> $elementType */
             if ($elementType::hasUris()) {
-                $elementIds = $refreshData->getElementIds($elementType);
-                $siteUris = array_merge($siteUris, SiteUriHelper::getElementSiteUris($elementIds));
+                $elementGroups = [];
+                foreach ($refreshData->getElementIds($elementType) as $elementId) {
+                    $siteIds = $refreshData->getElementSiteIds($elementType, $elementId);
+                    $siteIdsKey = $siteIds === null ? BaseDataModel::SITE_ID_ANY : implode(',', $siteIds);
+                    $elementGroups[$siteIdsKey]['siteIds'] = $siteIds;
+                    $elementGroups[$siteIdsKey]['elementIds'][] = $elementId;
+                }
+                foreach ($elementGroups as $elementGroup) {
+                    array_push($siteUris, ...SiteUriHelper::getElementSiteUris($elementGroup['elementIds'], $elementGroup['siteIds']));
+                }
             }
         }