+++ src/records/ElementFieldCacheRecord.php
@@ -10,6 +10,7 @@ use craft\db\ActiveRecord;
 /**
  * @property int $cacheId
  * @property int $elementId
+ * @property int $siteId
  * @property string $fieldInstanceUid
  *
  * @since 4.4.0